backend_example/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
openai==1.3.0  # OpenAI Python client
langchain==0.0.335  # For document processing and chat
tiktoken==0.5.1  # For token counting
PyPDF2==3.0.1  # For PDF processing
python-docx==1.1.0  # For Word document processing
tenacity==8.2.3  # For retry handling
